Implement controller with routes, refactor codebase
