+++ src/controllers/api/index.js
@@ -1,0 +1,3 @@
+export { default as AuthController } from './auth';
+export { default as UsersController } from './users';
+export { default as RedFlagsController } from './red-flags';
